feature: new jobs api

+++ protocol/responses.go
@@ -2,7 +2,7 @@ package protocol
 
 type errorResp struct {
 	Msg     string              `json:"message"`
-	Delay   int64               `json:"delay_seconds"`
+	Delay   int                 `json:"delay_seconds"`
 	Headers map[string][]string `json:"headers"`
 	Requeue bool                `json:"requeue"`
 }
